Add files via upload
Work on YCSB benchmarking of MySQL and MongoDB. Evaluation of results of the YCSB benchmarking test can be found here-further work is needed to complete this aspect of the assignment so further updates will come. Benchmarking data was collected using the terminal and the results were collected in individual text files and relevant content was combined into a CSV file manually.